Change from being a separate command, 'hg review', to wrapping 'hg push'.
This involved remapping the commandline options to not overlap with
any of the 'hg push' options.  I did this by having all options start
with 'r' (so --rr for reviewers, --rtitle for review title, etc).

The code now parses the review-related options, makes sure a valid
reviewer is specified, and then does the push followed by a
review-request in kiln.

This is only manually tested so far, but works for at least a simple
case.