The 'dest' arg to 'hg push' may not be an alias.  In that case, just take it.
_get_repo_to_push_to was always trying to look up the repo-name (the
argument to 'hg push') in the '[paths]' config in order to map the
repo alias to an actual url (or, I suppose, filesystem path).  But
one can legitimately just specify a url on the commandline. :-)  So if
we don't see a match in [paths], just use the dest argument as-is.