Add the ability to use kilnauth to get an authentication token for review.
It looks like kilnauth is meant to hook into mercurial.url.open, and
while I got that working, it didn't seem to set credentials the way I
wanted: it looks like the kiln API doesn't automatically take the
token from the fbToken header?

So instead I hack into kilnauth internals a bit in order to extract
the fbToken that it would put in the cookie, and adding it to our url
requests (token=<fbToken>).  Kiln is happy with that.

I also moved the authentication error-checking to the auth routines
where it belongs.